requirements.txt
Flask
gunicorn
sqlalchemy
pymysql
bcrypt
google-api-python-client
mysqlclient
python-dotenv